backend/pyproject.toml
[tool.poetry]
name = "sales-management"
version = "0.1.0"
description = ""
authors = ["Dan Kim <daniel@mysingle.io>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
alembic = "^1.14.0"
python-multipart = "^0.0.20"
emails = "^0.6"
jinja2 = "^3.1.5"
httpx = "^0.28.1"
tenacity = "^9.0.0"
passlib = "^1.7.4"
email-validator = "^2.2.0"
pydantic-settings = "^2.7.1"
sentry-sdk = "^2.19.2"
pyjwt = "^2.10.1"
pydantic = "^2.10.4"
psycopg = {extras = ["binary", "c"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
